modified how correspondences ratio is computed (icp 2D and 3D), also fixed build with OpenCV3+Cuda

+++ corelib/src/CameraRGB.cpp
@@ -115,7 +115,7 @@ unsigned int CameraImages::imagesCount() const
 {
 	if(_dir)
 	{
-		return _dir->getFileNames().size();
+		return (unsigned int)_dir->getFileNames().size();
 	}
 	return 0;
 }
